server/requirements.txt
Django==3.0.8
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.8.5
djangorestframework==3.11.0
django-cors-headers